Rename b to book and m to member in the creations methods.
Otherwise, because the m and b variables are also used in the dict comprehension, it can lead to errors (eg. modifying the wrong object).

This could have been resolved simply by placing the object creation after the usage of the list comprehension, but I find it clearer to have different variables names in here instead.